+++ Coimbra/Services/IApplicationService.cs
@@ -6,5 +6,11 @@ namespace Coimbra
     /// Responsible for the application lifetime cycle, meant to be used to fire Unity's callbacks events.
     /// </summary>
     [RequireImplementors]
-    public interface IApplicationService : IService { }
+    public interface IApplicationService : IService
+    {
+        /// <summary>
+        /// True when application is quitting.
+        /// </summary>
+        bool IsQuitting { get; }
+    }
 }